refactor: Refatoração rotinas de tratamento do retorno da API TCI Scorpio feat: Implementação de transmissão de audio no formato WAV através de socket
