+++ src/MIAerConnVS.vcxproj
@@ -175,12 +175,11 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClInclude Include="CRtspSession.h" />
-    <ClInclude Include="CStreamer.h" />
     <ClInclude Include="MIAerConnAudio.h" />
     <ClInclude Include="MIAerConnAudioCommon.h" />
     <ClInclude Include="MIAerConnCodes.hpp" />
     <ClInclude Include="ExternalCodes.h" />
+    <ClInclude Include="MIAerConnProcessApiResponse.h" />
     <ClInclude Include="MIAerLog.h" />
     <ClInclude Include="MIAerConnUtils.h" />
     <ClInclude Include="MoreEquipCtrlMsg.h" />
@@ -189,11 +188,10 @@
     <ClInclude Include="StdAfx.h" />
   </ItemGroup>
   <ItemGroup>
-    <ClCompile Include="CRtspSession.cpp" />
-    <ClCompile Include="CStreamer.cpp" />
     <ClCompile Include="ExternalCodes.cpp" />
     <ClCompile Include="MIAerConn.cpp" />
     <ClCompile Include="MIAerConnAudio.cpp" />
+    <ClCompile Include="MIAerConnProcessApiResponse.cpp" />
     <ClCompile Include="MIAerConnUtils.cpp" />
     <ClCompile Include="MIAerLog.cpp" />
     <ClCompile Include="StdAfx.cpp">
